siglip_service/requirements.txt
grpcio>=1.78.0
grpcio-health-checking>=1.78.0
protobuf>=5.29.0
torch>=2.2.0
transformers>=4.46.0
pillow>=10.2.0
